# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Playground",
    "version": "v1"
  },
  "paths": {
    "/Customer": {
      "get": {
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CustomersGetAllResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomersGetAllResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomersGetAllResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Customer"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomersCreateForm"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomersCreateForm"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CustomersCreateForm"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          }
        }
      }
    },
    "/Customer/{id}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CustomersGetByIdResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomersGetByIdResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomersGetByIdResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomersUpdateForm"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomersUpdateForm"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CustomersUpdateForm"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/Invoice": {
      "get": {
        "tags": [
          "Invoice"
        ],
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvoicesGetAllResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoicesGetAllResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoicesGetAllResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CustomersCreateForm": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CustomersGetAllModel": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CustomersGetAllResult": {
        "required": [
          "count",
          "items",
          "page",
          "pageCount",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "pageCount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "count": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "total": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomersGetAllModel"
            },
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "CustomersGetByIdModel": {
        "required": [
          "address",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "$ref": "#/components/schemas/DomainAddress"
          }
        },
        "additionalProperties": false
      },
      "CustomersGetByIdResult": {
        "required": [
          "customer"
        ],
        "type": "object",
        "properties": {
          "customer": {
            "$ref": "#/components/schemas/CustomersGetByIdModel"
          }
        },
        "additionalProperties": false
      },
      "CustomersUpdateForm": {
        "required": [
          "address",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "$ref": "#/components/schemas/DomainAddress"
          }
        },
        "additionalProperties": false
      },
      "DomainAddress": {
        "type": "object",
        "properties": {
          "line1": {
            "type": "string",
            "nullable": true
          },
          "line2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "zip": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InvoicesGetAllResult": {
        "required": [
          "count",
          "items",
          "page",
          "pageCount",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "pageCount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "count": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "total": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoicesModel"
            },
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "InvoicesModel": {
        "required": [
          "createdAt",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-05-23T13:48:11.3604364-04:00"
          }
        },
        "additionalProperties": false
      }
    }
  }
}